+++ 2019/7/2/circular_buffer.h
@@ -12,7 +12,7 @@ typedef circular_buf_t* cbuf_handle_t;
 /// Pass in a storage buffer and size, returns a circular buffer handle
 /// Requires: buffer is not NULL, size > 0
 /// Ensures: cbuf has been created and is returned in an empty state
-cbuf_handle_t circular_buf_init(int64_t* buffer, size_t size);
+cbuf_handle_t circular_buf_init(int* buffer, size_t size);
 
 /// Free a circular buffer structure
 /// Requires: cbuf is valid and created by circular_buf_init
@@ -26,17 +26,17 @@ void circular_buf_reset(cbuf_handle_t cbuf);
 /// Put version 1 continues to add data if the buffer is full
 /// Old data is overwritten
 /// Requires: cbuf is valid and created by circular_buf_init
-void circular_buf_put(cbuf_handle_t cbuf, int64_t data);
+void circular_buf_put(cbuf_handle_t cbuf, int data);
 
 /// Put Version 2 rejects new data if the buffer is full
 /// Requires: cbuf is valid and created by circular_buf_init
 /// Returns 0 on success, -1 if buffer is full
-int circular_buf_put2(cbuf_handle_t cbuf, int64_t data);
+int circular_buf_put2(cbuf_handle_t cbuf, int data);
 
 /// Retrieve a value from the buffer
 /// Requires: cbuf is valid and created by circular_buf_init
 /// Returns 0 on success, -1 if the buffer is empty
-int circular_buf_get(cbuf_handle_t cbuf, int64_t * data);
+int circular_buf_get(cbuf_handle_t cbuf, int * data);
 
 /// CHecks if the buffer is empty
 /// Requires: cbuf is valid and created by circular_buf_init
@@ -58,7 +58,7 @@ size_t circular_buf_capacity(cbuf_handle_t cbuf);
 /// Returns the current number of elements in the buffer
 size_t circular_buf_size(cbuf_handle_t cbuf);
 
-//TODO: int circular_buf_get_range(circular_buf_t cbuf, int64_t *data, size_t len);
-//TODO: int circular_buf_put_range(circular_buf_t cbuf, int64_t * data, size_t len);
+//TODO: int circular_buf_get_range(circular_buf_t cbuf, int *data, size_t len);
+//TODO: int circular_buf_put_range(circular_buf_t cbuf, int * data, size_t len);
 
 #endif //CIRCULAR_BUFFER_H_
